requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7
alembic>=1.12.1
pydantic>=2.9.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
python-multipart>=0.0.6
boto3>=1.29.7
redis>=5.0.1
websockets>=12.0
python-socketio>=5.10.0
geopy>=2.4.1

